add Command class, use it to implement usergroup
The Command class is a pretty simple way to add mock commands to the
shell. usergroup.py has examples for `getent`, `useradd`, and
`groupadd`.

It's not the most elegant thing in the world but hey, it works.

Note that all we're doing so far is parsing things into a bunch of
Namespace objects - code to actually generate useful stuff from those
objects will come later.